+++ dbknex.js
@@ -1,4 +1,4 @@
-const trace = require('./trace');
+const trace = require("./trace");
 
 const db = {
   host: process.env.DB_HOST,
@@ -8,14 +8,15 @@ const db = {
   database: process.env.DB_NAME,
 };
 
-const knex = require('knex')({
-  client: 'mysql',
+const knex = require("knex")({
+  client: "mysql",
   connection: db,
 });
 
-knex.raw('set names utf8;');
+knex.raw("set names utf8;");
+knex.raw("set global group_concat_max_len = 10000000;");
 
-knex.on('query', function (debugData) {
+knex.on("query", function (debugData) {
   trace.output(debugData);
 });