+++ routes/datasets.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { int } = require('../lib');
+const { int, Ok } = require('../lib');
 
 const datasets = require('../datasets.json');
 
@@ -9,9 +9,7 @@ router.get('/', (req, res, next) => {
   const data = datasets.filter(
     (dataset) => includeAll == 1 || !dataset.hidden || false
   );
-  res.status(200).json({
-    rs: data,
-  });
+  Ok(res, data);
 });
 
 module.exports = router;