+++ routes/races.js
@@ -6,9 +6,8 @@ const express = require('express');
 const router = express.Router();
 
 const knex = require('../dbknex');
-
-const { dsExists, stringOrDefault } = require('../lib');
-
+const { errorNotFound } = require('../errors');
+const { dsExists, stringOrDefault, Ok } = require('../lib');
 const trace = require('../trace');
 
 /**
@@ -18,7 +17,7 @@ const trace = require('../trace');
 router.get('/:ds', (req, res, next) => {
   const dbid = stringOrDefault(req.params.ds);
   if (!dsExists(dbid)) {
-    throw 'Unknown dataset';
+    throw { status:404, message:'Unknown dataset' };
   }
 
   where = '';
@@ -51,15 +50,13 @@ router.get('/:ds', (req, res, next) => {
     .where(knex.raw(where))
     .then(function (result) {
       if (result.length == 0) {
-        res.sendStatus(404);
+        errorNotFound(res);
       } else {
-        res.status(200).json({
-          rs: result,
-        });
+        Ok(res, result);
       }
     })
     .catch(function (error) {
-      if (error) throw error;
+      if (error) throw { message: error.message };
     });
 });
 
@@ -70,11 +67,11 @@ router.get('/:ds', (req, res, next) => {
 router.get('/:ds/:race', (req, res, next) => {
   const dbid = stringOrDefault(req.params.ds);
   if (!dsExists(dbid)) {
-    throw 'Unknown dataset';
+    throw { status:404, message:'Unknown dataset' };
   }
 
   const race = stringOrDefault(req.params.race);
-  if (race === '') throw 'Required URL argument not found';
+  if (race === '') throw { status:400, message:'Required URL argument not found' };
 
   const sql = [`select * from ${dbid}_races`, 'where race = ?'].join(' ');
 
@@ -86,15 +83,13 @@ router.get('/:ds/:race', (req, res, next) => {
     .where('race', race)
     .then(function (result) {
       if (result.length == 0) {
-        res.sendStatus(404);
+        errorNotFound(res);
       } else {
-        res.status(200).json({
-          rs: result,
-        });
+        Ok(res, result);
       }
     })
     .catch(function (error) {
-      if (error) throw error;
+      if (error) throw { message:error.message };
     });
 });