+++ routes/traits.js
@@ -6,8 +6,9 @@ const express = require('express');
 const router = express.Router();
 
 const knex = require('../dbknex');
+const { errorNotFound } = require('../errors');
 
-const { dsExists, stringOrDefault } = require('../lib');
+const { dsExists, stringOrDefault, Ok } = require('../lib');
 
 const trace = require('../trace');
 
@@ -20,17 +21,17 @@ const trace = require('../trace');
 router.get('/:ds', (req, res, next) => {
   const dbid = stringOrDefault(req.params.ds);
   if (!dsExists(dbid)) {
-    throw 'Unknown dataset';
+    throw { status:404, message:'Unknown dataset' };
   }
 
   const race = stringOrDefault(req.query.race);
   const profile = stringOrDefault(req.query.profile);
 
   if (race === '' && profile === '') {
-    throw 'Required URL argument not found';
+    throw { status:400, message:'Required URL argument not found' };
   }
   if (race !== '' && profile !== '') {
-    throw 'Too many URL arguments passed in';
+    throw { status:400, message:'Too many URL arguments passed in' };
   }
 
   let sql = '';
@@ -66,14 +67,12 @@ router.get('/:ds', (req, res, next) => {
 
   qb.then(function (result) {
     if (result.length == 0) {
-      res.sendStatus(404);
+      errorNotFound(res);
     } else {
-      res.status(200).json({
-        rs: result,
-      });
+      Ok(res, result);
     }
   }).catch(function (error) {
-    if (error) throw error;
+    if (error) throw { message: error.message };
   });
 });