+++ server.js
@@ -2,23 +2,23 @@
  * Express Server
  */
 
-const http = require('http');
-const dotenv = require('dotenv');
-const pkg = require('./package.json');
+const http = require("http");
+const dotenv = require("dotenv");
+const pkg = require("./package.json");
 
 const result = dotenv.config({
   silent: true,
   debug: process.env.DEBUG,
 });
 if (!result.error) {
-  console.log('Parsed environment... ', result.parsed);
+  console.log("Parsed environment... ", result.parsed);
 }
 
 // Defaults to alwaysdata IP & PORT values
-const ip = process.env.IP || '0.0.0.0';
+const ip = process.env.IP || "0.0.0.0";
 const port = process.env.PORT || 8100;
 
-const app = require('./app');
+const app = require("./app");
 const server = http.createServer(app);
 
 server.listen(port, ip, () => {